Implenting file_toolbar and styling file browser
Work is being started to add interactive elements to the project such as
being able to add new rows to a table so that the user can create files
and there are now plans to add the ability to upload files as well as
download files.
There is a planned refactor for file_browser soon but that sounds like a
weekend job.